декомпозирует компонент MainPage
