+++ project/src/components/header-nav-logged/header-nav-logged.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { fetchInfoAuthAction, logoutAction } from '../../store/api-actions';
+import { logoutAction } from '../../store/api-actions';
 
 function HeaderNavLogged(): JSX.Element {
   const email = useAppSelector((state) => state.userInfo?.email);
@@ -20,7 +20,6 @@ function HeaderNavLogged(): JSX.Element {
           onClick={(evt) => {
             evt.preventDefault();
             dispatch(logoutAction());
-            dispatch(fetchInfoAuthAction());
           }}
           className="header__nav-link"
           to="/"
